init rewrite
